fix: Footer component
